LayoutTests/http/tests/w3c/webperf/submission/resource-timing/html/test_resource_frame_initiator_type-expected.txt

Description

This test validates that the frame initiator type is represented.



PASS /w3c/webperf/resources/blank_page_green.htm is expected to have initiatorType frame 
